src/requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_cors
flask_admin
marshmallow
marshmallow-sqlalchemy
pymysql
cryptography
pandas
numpy
scikit-learn==1.7.2
plotly
joblib
requests
gunicorn
matplotlib
